requirements_gpu.txt
tensorflow-gpu==2.6.0
Keras==2.6.0
numpy==1.19.5
h5py==3.1.0
lxml==4.3.4
mrcfile==1.1.2
scikit-learn==0.22.2.post1
scikit-image==0.15.0
matplotlib==3.1.0
PyQt5==5.13.2
pyqtgraph==0.10.0
openpyxl==3.0.3